During testing we observed that the MaxNewSize i.e. par new size was set as very less as compared to old gen size. As an example, oldgen size was set as 91GB while the par new size was set as 1.5 GB.
Actually when UseConcMarkSweepGC property is set, jdk sets the new size as a very low number. It has been discussed in the following open jdk issue.

https://bugs.openjdk.java.net/browse/JDK-8153578

We need the new gen size to be on the higher side as there are too many temporary objects being created. Setting it as a 1/4 of the JVM size for now.