Executorcatalog (#146)
Changes proposed in this pull request

The cached batch size is saved in the hive catalog per table and this hive meta data is available on executor. The hive meta data is fetched once and then is kept in the container. ExternalTableMetaData is the class that holds the hive meta data. All hive meta data is invalidated when a new container is created. This is brute force but currently when hive meta data is updated, no events are received on the executor.

Removed the unused functions isUsedForUserReplicatedTable and setUsedForUserReplicatedTable.
Patch testing

Clean precheckin. Will be adding a test to verify the hive meta data on executors in embedded and connector modes.
Other PRs

SnappyDataInc/snappydata#500