Flip the ByteBuffer after read/write in the new TCompactProtocolOpt/TBinaryProtocolOpt
Thrift code always assumes a heap array and uses it directly so never needs to read/write ByteBuffer
and hence it does not have the code to flip the buffer after read/write to channel.
The new optimized implementations need to do that explicitly.

Allow for passing in connection-level lob-chunk-size property which will be honoured for
all Statement executions on that connection.

Some other logging changes, minor optimizations.