[SNAP-1960] correct setting of RUNNING status
PR#182 for SNAP-1517 added an "endWaiting" flag which was false in normal launch sequence
and reset by "endNotifyWaiting" callback. This, however, caused SNAP-1893 because "endNotify"
would update only for the case when previous state was RUNNING. So PR#276 added another
call to set launch status to RUNNING when internal server status is set so.

The above changes cause premature setting of RUNNING which happens at the end of GemFireStore
boot before even DDL replay is complete (which itself is another issue). Secondly the launch
sequence also starts up additional services like thrift server, so RUNNING should be set
only after that is complete and not just at the end of boot sequence. The original issue
SNAP-1517 could have been solved simply by setting RUNNING status for the normal launch
sequence when it is WAITING and without introducing any new calls, but this PR also
corrects the internal status and switches status correctly as required.

- pass the status to set from WAITING to CacheServerLauncher.running instead of a boolean
- the full launch sequence will call this with RUNNING at the end while the callback
  of "endNotifyWaiting" will set it to the previous status (STARTING)

  So the launch status will change as: STARTING -> WAITING -> STARTING -> RUNNING,
  which is the correct way because once WAITING it done then GII etc can take a long time

- remove "notifyRunning" call from GemFireStore boot rather only at the end of
  DDL replay in FabricDatabase.postCreate to correct the internal state too

  The failure in BackwardCompatabilityDUnit was due to above where client would try to
  connect to servers much before the thrift servers have started (during DDL replay)

- also corrected the BackwardCompatabilityDUnit test to wait for locator start before
  starting the servers rather than starting locator+servers in parallel (that cause
      occasional failure in servers due to locator unavailability)