Fix logging to add threadID for all threads
Currently the MDC mechanism used for adding threadID is thread-local
set in GFToSlf4jBridge to threads using Spark Logging/slf4j/log4j
directly don't get the threadID (appears as "Thread..<tid=>").

This changes to use a more general way of updating the PatternLayout
that updates the threadName in the LoggingEvent itself to append the
threadID in the above format. However, this will not work if user
has configured AsyncAppender which will run in a separate dedicated
thread so an alternative using ThreadIdAppender has also been provided
that needs to be applied before AsyncAppender which will update LoggingEvent.