fix a rare failure in CreateTableTest.testSerializedException_43160
Due to expected behaviour of replicated region checking for lastModifiedTime to determine
if value is a repeat insert in a putAll when value bytes are the same.
In this case the duplicate inserts may go in with the same lastModifiedTime so try
again after a sleep to force the lastModifiedTime to be different from already inserted entries.