Fix switching between disk and remote iterators
If some buckets are remote and some local on disk that are intermixed,
then the current PR iterator would switch to remote but not switch back
to original disk iterator rather create a new map entries iterator
(which is incorrect to begin with).

Now store the disk iterator being used and switch back to it when a
local bucket is seen after a remote one. This disk iterator is the
one initialized at the end for on-disk entries after all in-memory
entries have been read. Also ensure that it is cleaned up in close
in case the last bucket iterator was a remote one.