Fixed lock timeout issues seen in Snappy precheckin
Reason was that catalog inconsistency detection will create a connection that marked the
connection ID as CHILD_NOT_CACHEABLE. This caused no connection to be tracked on remote
nodes to GfxdDDLFinishMessage would skip any processing.

This change marks the connection as a proper one with UINITIALIZED connection ID that
will get assigned with a unique value in the EmbedConnection constructor.
Also added few logging changes to track such cases better.