Set load-balance=false on servers by default
Previous set of changes checked if the host name specified in JDBC URL is a locator or server type.
This matching would fail for cases where the host in JDBC URL is different from the hostname-for-clients
setting on the servers (e.g. when using internal IP in an AWS cluster that has public IP in hostname-for-clients).

This change adds a new flag "isCurrent" in HostAddress that marks the connected server among the received list.
The flag is then used to determine whether that member is a server or locator and load-balance is set accordingly.