Added password encryption
Provide password encryption including for SSL keystores, enabled by
setting "encrypted-passwords=true". The password field is then assumed
to be the target name to be looked up in system credential manager to
obtain the password.

On Windows this uses the standard CredRead() API with type as
CRED_TYPE_GENERIC so user must add the password under that heading.

On Mac OSX the "security" tool is used to lookup the password for the
given value as provided in the password field.

On Linux and other systems, the "secret-tool" is used to lookup the
password with the password field split on the first ':' to obtain the
attribute and its value.