changed debug compiler flag to "-g" from "-g3"
Newer gdb+gcc releases are terribly slow with "-g3" at least with the C++ driver
for some unknown reason so switching it to "-g" for both gradle and eclipse.
The "-g3" supposedly adds more information for macros which might be causing
trouble with the complicated boost macros. Not important anyway.