[SNAP-3332] pass back scale/precision array in PreparedStatement execution  (#567)
The prepare phase of PreparedStatement returns the scale/precision of parameters for routed
queries. This should be passed back during execution so that the current set of parameters
are set on the SnappySession rather than those set by the last prepare (which can be different)

Also made the PVS info array size uniform to be three ints per type removing the unused
(and incorrect) fourth field which is supposed to have nullability information.